Convert VirtoLogo to typescript
