+++ src/SvgIcons/VirtoLogo/VirtoLogo.types.ts
@@ -1,0 +1,3 @@
+import { SvgIconProps } from "../SvgIcons.types";
+
+export type VirtoLogoProps = SvgIconProps;
