+++ src/SvgIcons/VirtoLogo/index.ts
@@ -1,13 +1,2 @@
-import { SxProps, Theme } from "@mui/material/styles";
-
-import { SvgIconProps } from "../SvgIcons.types";
-import VirtoLogo from "./VirtoLogo";
-
-export type VirtoLogoProps = SvgIconProps & {
-  /**
-   * Styling applied to the component
-   */
-  sx?: SxProps<Theme>;
-};
-
-export default VirtoLogo as React.FC<VirtoLogoProps>;
+export { VirtoLogo } from "./VirtoLogo";
+export type { VirtoLogoProps } from "./VirtoLogo.types";
